move project visited to get project info api
