+++ server/controller/src/main/java/ai/starwhale/mlops/configuration/security/SecurityConfiguration.java
@@ -84,8 +84,6 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
     @Resource
     private ContentCachingFilter contentCachingFilter;
 
-    @Resource
-    private VisitedFilter visitedFilter;
 
     public SecurityConfiguration() {
         super();
@@ -141,7 +139,6 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
                         JwtLoginFilter.class)
                 .addFilterBefore(projectDetectionFilter, JwtTokenFilter.class)
                 .addFilterBefore(contentCachingFilter, ProjectDetectionFilter.class)
-                .addFilterAfter(visitedFilter, JwtTokenFilter.class)
         ;
     }
